Remove unprogram card confirmation modal
